Importing of stage one data matrix to PReMiuM
The stage one matrix is passed into profRegr() as extra columns to the
data parameter, and the names of the column passed in as a character
vector in outcome.

This code does not account for anything other than discrete models at
the moment.

Move between different values for Y are made, but they don't account for
the model at all - all moves are accepted with probability 0.36